+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Long Break never happening (because of bad parsing of preferences)
+
 ## [0.0.4] - 2022-08-31
 ### Changed
 - allow to edit preferences only when breaks are stopped